Refactor ResponseObject as inner class
